--- a/paddle/phi/kernels/funcs/dense_tensor_iterator.h
+++ b/paddle/phi/kernels/funcs/dense_tensor_iterator.h
@@ -19,8 +19,6 @@
 #include "paddle/common/ddim.h"
 #include "paddle/phi/core/dense_tensor.h"
 
-#include "glog/logging.h"
-
 namespace phi {
 
 class DenseTensorIteratorConfig;