<!-- TemplateReference: https://github.com/PaddlePaddle/Paddle/wiki/PULL-REQUEST-TEMPLATE--REFERENCE -->
<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->

### PR Category
<!-- One of [ User Experience | Execute Infrastructure | Operator Mechanism | CINN | Custom Device | Performance Optimization | Distributed Strategy | Parameter Server | Communication Library | Auto Parallel | Inference | Environment Adaptation ] -->
Execute Infrastructure

### PR Types
<!-- One of [ New features | Bug fixes | Improvements | Performance | BC Breaking | Deprecations | Docs | Devs | Not User Facing | Security | Others ] -->
Others

### Description
<!-- Describe what you’ve done -->
pcard-67164

# PR简介：基于DenseTensorIterator的算子计算机制优化

## 一、简介

本PR以Paddle DenseTensor数据结构为基础，设计并实现了基于**DenseTensorIterator**的算子计算机制，用于在处理非连续输入时获得性能提升/显存占用降低以及正确性生态对齐。

## 二、整体设计

<img width="2426" height="2975" alt="流程图-202508181405 (1)" src="https://github.com/user-attachments/assets/9c85895d-5aa7-4d39-9d63-c59f75b84c41" />

## 三、实现方法

本PR主要实现了设计中的**Paddle TensorIterator**以及其相关的**通用CUDA Stride Kernel**部分，通过以上实现确保了elementwise_add算子性能及正确性提升，下面由上到下讲解具体实现。

### 1. Strided Kernel

该部分需要注册并实现一个GPU上的stride kernel。

- **实现位置**: `paddle/phi/kernels/stride/elementwise_kernel.cu`
- **实现方法**: 使用`PD_REGISTER_KERNEL`进行定义，并在`elementwise_kernel.cu`中实现
- **函数命名规则**: `XXStrideKernel`
- **扩展机制**: 为方便后续扩量生成StrideKernel，已使用宏语言编程方式做了处理，后续只需要通过例如`DEFINE_CUDA_BINARY_ELEMENTWISE_STRIDE_OP(Add)`的方式生成函数即可

### 2. Risk Control (Flag)

该部分需要实现一个用于风险管理的Flag，预防新增功能不稳定带来的正确性以及性能影响。

- **实现位置**: `paddle/common/flags.cc`
- **Flag名称**: `FLAGS_use_stride_compute_kernel`
- **使用方式**: 
 - 当使用`FLAGS_use_stride_compute_kernel=1`运行Paddle程序时，Kernel会分派到新的实现方法，直接处理非连续输入运算
 - 反之则会回退到之前版本，先将输入变连续再进行运算

### 3. Paddle TensorIterator (DenseTensorIterator)

该部分需要实现一个对任意张量进行运算前预处理的类，目前命名为**DenseTensorIterator**。

- **实现位置**: 
 - `paddle/phi/kernels/funcs/dense_tensor_iterator.cc`
 - `paddle/phi/kernels/funcs/dense_tensor_iterator.h`
- **主要功能**: DenseTensorIterator保证了张量进入通用CUDA kernel前拥有最佳的shape/stride以供后续运算，同时保证了输出张量的连续性与生态对齐

### 4. CUDA strided_kernel_impl

该部分需要实现一个通用的CUDA kernel，用于处理DenseTensorIterator给出的shape和stride，并按照Functor的定义完成运算。

- **实现位置**: `paddle/phi/kernels/stride/elementwise_kernel.cu`
- **实现方法**: 使用Slice优化时引入的`OffsetCalculator`进行index计算，之后复用KP中的Functor相关实现进行运算
- **编译优化**: 为了保证不带来过多的编译体积增加，本PR去除了KP相关的模版，只保留type以及Functor相关模版，保证了编译体积可控

## 四、相关影响

### 1. 正确性影响

当`FLAGS_use_stride_compute_kernel=1`时，elementwise_add将按照如下规则进行输出处理：

对于两个参与二元Elementwise操作的操作数，一个必要的前提是需要将两个操作数能广播到一个共同的尺寸上，我们定义这个尺寸为**广播尺寸**。

- 如果两个操作数的尺寸都为广播尺寸，那么输出的stride由第一个操作数决定（按顺序决定）
- 如果其中一个操作数需要广播到广播尺寸，那么输出的stride由与广播尺寸相同的操作数决定（与顺序无关）

当`FLAGS_use_stride_compute_kernel=0`时，elementwise_add将全输出连续张量。

### 2. 性能影响

#### Transpose + Add 性能测试

| Shape | Before (μs) | After (μs) | Improvement |
|-------|------------|------------|-------------|
| 64×108×12288 | 2262.79 | 1243.12 | **81.9% faster** |

#### As_strided + Add 性能测试

| Shape | Before (μs) | After (μs) | Improvement |
|-------|------------|------------|-------------|
| 54×32×12288 | 651.61 | 451.13 | **44.3% faster** |

## 五、风险管理

为保证`FLAGS_use_stride_compute_kernel=1`后正确性可以保证，对elementwise_add新增了单测。

- **单测位置**: `test/legacy_test/test_elementwise_add_op.py`
- **覆盖场景**: 单测覆盖了transpose/as_strided、广播、0dim、0size等多种情况，最大限度保证了正确性